requirements.txt
adjustText==1.2.0
appdirs==1.4.4
attrs==22.2.0
cattrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
contourpy==1.1.0
cycler==0.11.0
exceptiongroup==1.1.0
fonttools==4.40.0
highlight-text==0.2
idna==3.4
inflect==6.0.2
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.8.2
mplsoccer==1.2.4
nose2==0.12.0
numpy==1.24.1
packaging==23.1
pandas==2.1.3
Pillow==9.5.0
pip==23.3
pydantic==1.10.4
pyparsing==3.1.0
pystatsbomb==0.1.0
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests-cache==0.9.8
requests==2.31.0
scipy==1.11.2
seaborn==0.13
setuptools==60.2.0
scikit-learn==1.6.1
six==1.16.0
statsbombpy==1.13
shapely == 2.0.5
shiny==1.3.0
url-normalize==1.4.3
urllib3==3.2.0
wheel==0.37.1
pypalettes == 0.1.0
typing-extensions==4.10.0rc1
thefuzz == 0.22.1
rsconnect_python==1.24.0

